+++ Assets/MenuUi/Scripts/DailyQuest/DailyTaskProgressPopup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4719449810754fc49a3a95749efc3992
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
